Add NSDictionary-to-JSON conversion benchmark. An ideal JSON library will maintain data as an NSDictionary internally rather than force a conversion to a Swift dictionary during init.
On iPhone 5C, JASON runs in 0.002 sec, SwiftyJSON runs in 0.030 sec.

JSON is typically originates as an NSDictionary, and converting
a large NSDictionary to a Swift dictionary is slow (due to mem copy and casting of
every item in the dict).

Type conversions to Swift native types should happen only on key access, thus
amortizing the cost (or avoiding the bulk conversion cost in the case where
there is no consecutive loop access of every key in the JSON data).